tidy up packages and redo tests with mocks
