+++ src/main/java/kata/Controller/SkuController.java
@@ -1,10 +1,13 @@
 package kata.Controller;
 
-import kata.Model.repo.ItemRepo;
-import kata.Model.sku.Sku;
+import kata.Repository.ItemRepo;
+import kata.Model.Sku;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.net.URISyntaxException;
 
 @RestController
 @RequestMapping("/skus")
@@ -25,10 +28,14 @@ public class SkuController {
         }
     }
 
-    @PostMapping("/post")
-    public Sku addNewSku(@RequestBody Sku newSku) {
+    @PostMapping()
+    public ResponseEntity addNewSku(@RequestBody Sku newSku) throws URISyntaxException {
+        if(itemRepo.getSku(newSku.getNameOfProduct()) != null){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(String.format("SKU already exists", newSku.getNameOfProduct()));
+        }
+
         itemRepo.addSku(newSku);
-        return newSku;
+        return ResponseEntity.created(new URI("skus/" + newSku.getNameOfProduct())).body(newSku);
     }
 }
