+++ src/main/java/kata/Model/Offer.java
@@ -1,0 +1,21 @@
+package kata.Model;
+
+public class Offer {
+    private int quantityOfProduct;
+    private int priceInPence;
+
+    public Offer(int quantityOfProduct, int priceInPence) {
+        this.quantityOfProduct = quantityOfProduct;
+        this.priceInPence = priceInPence;
+    }
+
+    public int getQuantityOfProduct() {
+        return quantityOfProduct;
+    }
+
+    public int getPriceInPence() {
+        return priceInPence;
+    }
+
+    }
+
