--- src/main/java/kata/Model/offer/Offer.java
@@ -1,21 +1,0 @@
-package kata.Model.offer;
-
-public class Offer {
-    private int quantityOfProduct;
-    private int priceInPence;
-
-    public Offer(int quantityOfProduct, int priceInPence) {
-        this.quantityOfProduct = quantityOfProduct;
-        this.priceInPence = priceInPence;
-    }
-
-    public int getQuantityOfProduct() {
-        return quantityOfProduct;
-    }
-
-    public int getPriceInPence() {
-        return priceInPence;
-    }
-
-    }
-
