--- src/test/java/kata/Model/TestRunner.java
@@ -1,17 +1,0 @@
-package kata.Model;
-import kata.Model.checkout.CheckoutServiceTest;
-import org.junit.runner.JUnitCore;
-import org.junit.runner.Result;
-import org.junit.runner.notification.Failure;
-
-
-public class TestRunner {
-    public static void main(String[] args) {
-        Result result = JUnitCore.runClasses(CheckoutServiceTest.class);
-
-        for (Failure failure : result.getFailures()) {
-            System.out.println(failure.toString());
-        }
-        System.out.println(result.wasSuccessful());
-    }
-}
