--- src/test/java/kata/Model/repo/OfferRepoTest.java
@@ -1,18 +1,0 @@
-package kata.Model.repo;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-import static org.mockito.Mockito.verify;
-
-public class OfferRepoTest {
-
-    @Test
-    public void getOffers() {
-
-    }
-
-    @Test
-    public void getOffer() {
-    }
-}
